# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ExchangeType": {
				"description": "Enum representing supported exchanges",
				"enum": [
					"BINANCE",
					"BYBIT"
				],
				"type": "string"
			},
...
			"TradingPair": {
				"properties": {
					"buyPrice": {
						"type": "number",
						"format": "double"
					},
					"sellPrice": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"buyOrderId": {
						"type": "number",
						"format": "double"
					},
					"sellOrderId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"buyPrice",
					"sellPrice",
					"quantity",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TradingPair-Array_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TradingPair"
						},
						"type": "array",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewTradingPair": {
				"properties": {
					"buyPrice": {
						"type": "number",
						"format": "double"
					},
					"sellPrice": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"buyOrderId": {
						"type": "number",
						"format": "double"
					},
					"sellOrderId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"buyPrice",
					"sellPrice",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TradingPair_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/TradingPair",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_NewTradingPair_": {
				"properties": {
					"buyPrice": {
						"type": "number",
						"format": "double"
					},
					"sellPrice": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"buyOrderId": {
						"type": "number",
						"format": "double"
					},
					"sellOrderId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"ApiResponse_null_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Ticker": {
				"description": "Unified ticker interface for all exchanges",
				"properties": {
					"last": {
						"type": "number",
						"format": "double"
					},
					"changePercentage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"last",
					"changePercentage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.Ticker_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/Ticker"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ApiResponse_Record_string.Ticker__": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.Ticker_",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Product": {
				"description": "Unified product model for exchange products/trading pairs",
				"properties": {
					"currencyPair": {
						"type": "string"
					},
					"minQuantity": {
						"type": "number",
						"format": "double"
					},
					"minTotal": {
						"type": "number",
						"format": "double"
					},
					"pricePrecision": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"currencyPair",
					"minQuantity",
					"minTotal",
					"pricePrecision"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Product-Array_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SymbolInfo": {
				"properties": {
					"baseAsset": {
						"type": "string"
					},
					"quoteAsset": {
						"type": "string"
					},
					"stepSize": {
						"type": "string"
					},
					"tickSize": {
						"type": "string"
					},
					"minNotional": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"baseAsset",
					"quoteAsset",
					"stepSize",
					"tickSize",
					"minNotional"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_SymbolInfo_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/SymbolInfo",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_BinanceBotOrder-Array_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/BinanceBotOrder"
						},
						"type": "array",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BotConfig": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"BotTradingPair": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"Bot": {
				"properties": {
					"config": {
						"$ref": "#/components/schemas/BotConfig"
					},
					"pairs": {
						"items": {
							"$ref": "#/components/schemas/BotTradingPair"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"config",
					"pairs",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Bot_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/Bot",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Bot-Array_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Bot"
						},
						"type": "array",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_OrderStatistics_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/OrderStatistics",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_BinanceBotOrder_": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"$ref": "#/components/schemas/BinanceBotOrder",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__ordersDeleted-number--testnetOrdersDeleted-number--filledOrdersQueueDeleted-number--testnetFilledOrdersQueueDeleted-number--botDeleted-boolean__": {
				"description": "Base response interface for all API responses",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"error": {
						"type": "string",
						"description": "Optional error message in case of failure"
					},
					"data": {
						"properties": {
							"botDeleted": {
								"type": "boolean"
							},
							"testnetFilledOrdersQueueDeleted": {
								"type": "number",
								"format": "double"
							},
							"filledOrdersQueueDeleted": {
								"type": "number",
								"format": "double"
							},
							"testnetOrdersDeleted": {
								"type": "number",
								"format": "double"
							},
							"ordersDeleted": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"botDeleted",
							"testnetFilledOrdersQueueDeleted",
							"filledOrdersQueueDeleted",
							"testnetOrdersDeleted",
							"ordersDeleted"
						],
						"type": "object",
						"description": "Optional data payload in case of success"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "spot-trading-hub",
		"version": "1.0.0",
		"description": "A Node.js application for managing crypto exchange spot orders",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/trading-pairs": {
			"get": {
				"operationId": "GetAllTradingPairs",
				"responses": {
					"200": {
						"description": "Trading pairs retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TradingPair-Array_"
								}
							}
						}
					}
				},
				"description": "Get all trading pairs",
				"tags": [
					"Trading Pairs"
				],
				"security": [],
				"parameters": [
					{
						"description": "Bot ID to filter trading pairs",
						"in": "query",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTradingPairs",
				"responses": {
					"201": {
						"description": "Trading pairs created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TradingPair-Array_"
								}
							}
						}
					}
				},
				"description": "Create multiple trading pairs",
				"tags": [
					"Trading Pairs"
				],
				"security": [],
				"parameters": [
					{
						"description": "Bot ID for the trading pairs",
						"in": "query",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Array of trading pairs to create (without id)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/NewTradingPair"
								},
								"type": "array",
								"description": "Array of trading pairs to create (without id)"
							}
						}
					}
				}
			}
		},
		"/trading-pairs/{id}": {
			"put": {
				"operationId": "UpdateTradingPair",
				"responses": {
					"200": {
						"description": "Trading pair updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TradingPair_"
								}
							}
						}
					}
				},
				"description": "Update a trading pair by ID",
				"tags": [
					"Trading Pairs"
				],
				"security": [],
				"parameters": [
					{
						"description": "Trading pair ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Partial trading pair data to update",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_NewTradingPair_",
								"description": "Partial trading pair data to update"
							}
						}
					}
				}
			}
		},
...
		"/exchange/{exchange}/trader-tickers": {
			"get": {
				"operationId": "GetTraderTickers",
				"responses": {
					"200": {
						"description": "Tickers retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Record_string.Ticker__"
								}
							}
						}
					}
				},
				"description": "Get tickers from a specific exchange",
				"tags": [
					"Exchange"
				],
				"security": [],
				"parameters": [
					{
						"description": "The exchange to get tickers from",
						"in": "path",
						"name": "exchange",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ExchangeType"
						}
					}
				]
			}
		},
		"/exchange/{exchange}/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Products retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Product-Array_"
								}
							}
						}
					}
				},
				"description": "Get available trading products/pairs from a specific exchange",
				"tags": [
					"Exchange"
				],
				"security": [],
				"parameters": [
					{
						"description": "The exchange to get products from",
						"in": "path",
						"name": "exchange",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ExchangeType"
						}
					}
				]
			}
		},
		"/exchange/{exchange}/symbol/{symbol}": {
			"get": {
				"operationId": "GetSymbolInfo",
				"responses": {
					"200": {
						"description": "Symbol information retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_SymbolInfo_"
								}
							}
						}
					}
				},
				"description": "Get symbol information from a specific exchange",
				"tags": [
					"Exchange"
				],
				"security": [],
				"parameters": [
					{
						"description": "The trading pair symbol",
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The exchange to get symbol info from",
						"in": "path",
						"name": "exchange",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ExchangeType"
						}
					}
				]
			}
		},
		"/binance-bot/{botId}/orders": {
			"get": {
				"operationId": "GetBotOrders",
				"responses": {
					"200": {
						"description": "Bot orders",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BinanceBotOrder-Array_"
								}
							}
						}
					}
				},
				"description": "Get all orders for a specific bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/binance-bot/{botId}": {
			"get": {
				"operationId": "GetBot",
				"responses": {
					"200": {
						"description": "Bot",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Bot_"
								}
							}
						}
					}
				},
				"description": "Get configuration for a specific bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/binance-bot/{botId}/config": {
			"put": {
				"operationId": "UpdateBotConfig",
				"responses": {
					"200": {
						"description": "Bot configuration updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Update configuration for a specific bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The bot configuration",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BotConfig",
								"description": "The bot configuration"
							}
						}
					}
				}
			}
		},
		"/binance-bot/{botId}/pairs": {
			"put": {
				"operationId": "UpdateBotPairs",
				"responses": {
					"200": {
						"description": "Bot pairs updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Update trading pairs for a specific bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The bot trading pairs",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/BotTradingPair"
								},
								"type": "array",
								"description": "The bot trading pairs"
							}
						}
					}
				}
			}
		},
		"/binance-bot": {
			"post": {
				"operationId": "CreateBot",
				"responses": {
					"201": {
						"description": "Bot configuration created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Bot_"
								}
							}
						}
					}
				},
				"description": "Create configuration for a specific bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The bot configuration",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BotConfig",
								"description": "The bot configuration"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllBots",
				"responses": {
					"200": {
						"description": "All bots",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Bot-Array_"
								}
							}
						}
					}
				},
				"description": "Get all bots",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/binance-bot/{botId}/statistics": {
			"get": {
				"operationId": "GetBotStatistics",
				"responses": {
					"200": {
						"description": "Bot order statistics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_OrderStatistics_"
								}
							}
						}
					}
				},
				"description": "Get statistics for a specific bot's orders",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/binance-bot/{botId}/filled-order": {
			"post": {
				"operationId": "AddBotFilledOrder",
				"responses": {
					"200": {
						"description": "Order",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BinanceBotOrder_"
								}
							}
						}
					}
				},
				"description": "Add a Binance order to the bot's orders",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The order request containing orderId and symbol",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"symbol": {
										"type": "string"
									},
									"orderId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"symbol",
									"orderId"
								],
								"type": "object",
								"description": "The order request containing orderId and symbol"
							}
						}
					}
				}
			}
		},
		"/binance-bot/cleanup/testnet/{botId}": {
			"delete": {
				"operationId": "CleanupTestnetBot",
				"responses": {
					"200": {
						"description": "Testnet bot data cleaned up",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__ordersDeleted-number--testnetOrdersDeleted-number--filledOrdersQueueDeleted-number--testnetFilledOrdersQueueDeleted-number--botDeleted-boolean__"
								}
							}
						}
					}
				},
				"description": "Cleanup all data related to a testnet bot",
				"tags": [
					"Binance Bot"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the bot to cleanup",
						"in": "path",
						"name": "botId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	],
...
}